Update how the Microsoft Entra Registered section
Update the Microsoft Entra Registered section.

It's not clear what "Not Applicable" really means. We need more clarification on that, as it has been causing a lot of doubts among SEs, including the most senior ones and even SEE. Please engage with me at v-frusso@microsoft.com if you have any questions regarding this, as I believe it is very important to proceed with these changes. Should have a disclaimer explainning why